Create README.md
This is a Math project that contains various functions including class that represent a single Monom (ax^b) that has two variables coefficient and power, and another class that represent Polynom which is a collection of Monoms that are saved in ArrayList which is sorted from highest power to the lowest. Each of those classes has a different method such as adding two Monoms with the same power ,finding a derivative for a Polynom , multiply , equals , root , area etc.